document_service/requirements.txt
flask
requests
redis
rq
pytest
pdfminer.six